frontend/requirements.txt
langgraph
langchain
langchain-huggingface
langchain-community
python-dotenv
fastapi
uvicorn
google-search-results
pydantic
PyPDF2
python-multipart
